second task: docker, viper, goose, logger

+++ deploy/0001_init.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS payments (
     id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     payment_id INTEGER UNIQUE NOT NULL,
-    amount INTEGER
+    amount BIGINT
 );
 
 -- +goose Down
